…ingNext

The optimization added here is only applicable when the sstable is not exhausted. The existing IteratorSeqSeekPrefixGENotFound benchmark only improves for the case where bloom=false, since bloom=true will fail in the bloom filter match.

```
name                                                                                     old time/op    new time/op    delta
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=false/with-tombstone=false-10     262ns ± 1%     263ns ± 1%     ~     (p=0.151 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=false/with-tombstone=true-10      205ns ± 1%     207ns ± 0%   +0.86%  (p=0.016 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=true/with-tombstone=false-10      511ns ± 1%     537ns ± 2%   +5.12%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=true/with-tombstone=true-10       295ns ± 2%     304ns ± 2%   +3.12%  (p=0.016 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=false/with-tombstone=false-10     1.89µs ± 0%    0.30µs ± 1%  -84.41%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=false/with-tombstone=true-10       575ns ± 0%     213ns ± 1%  -62.92%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=true/with-tombstone=false-10       878ns ± 0%     540ns ± 3%  -38.55%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=true/with-tombstone=true-10        660ns ± 0%     308ns ± 5%  -53.32%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=false/with-tombstone=false-10     272ns ± 0%     275ns ± 1%   +1.11%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=false/with-tombstone=true-10      217ns ± 0%     220ns ± 1%   +1.60%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=true/with-tombstone=false-10      522ns ± 3%     562ns ± 6%   +7.65%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=true/with-tombstone=true-10       310ns ± 1%     311ns ± 1%     ~     (p=0.556 n=4+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=false/with-tombstone=false-10     1.89µs ± 1%    0.31µs ± 1%  -83.77%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=false/with-tombstone=true-10       581ns ± 1%     226ns ± 1%  -61.17%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=true/with-tombstone=false-10       888ns ± 0%     545ns ± 6%  -38.68%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=true/with-tombstone=true-10        677ns ± 1%     312ns ± 1%  -53.98%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=false/with-tombstone=false-10     299ns ± 0%     296ns ± 1%   -1.11%  (p=0.032 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=false/with-tombstone=true-10      244ns ± 0%     242ns ± 0%   -1.15%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=true/with-tombstone=false-10      568ns ± 5%     560ns ± 0%     ~     (p=0.690 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=true/with-tombstone=true-10       346ns ± 3%     333ns ± 0%   -3.68%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=false/with-tombstone=false-10     1.93µs ± 1%    0.33µs ± 0%  -83.14%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=false/with-tombstone=true-10       607ns ± 0%     248ns ± 1%  -59.21%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=true/with-tombstone=false-10       951ns ± 6%     558ns ± 0%  -41.32%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=true/with-tombstone=true-10        717ns ± 8%     339ns ± 2%  -52.76%  (p=0.008 n=5+5)

name                                                                                        old time/op  new time/op  delta
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=false/with-tombstone=false-10   216ns ± 0%   217ns ± 0%     ~     (p=0.151 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=false/with-tombstone=true-10    178ns ± 1%   175ns ± 0%   -1.43%  (p=0.016 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=true/with-tombstone=false-10    430ns ± 5%   437ns ± 2%     ~     (p=0.135 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=true/with-tombstone=true-10     106ns ± 1%   105ns ± 1%     ~     (p=0.373 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=false/with-tombstone=false-10   1.20µs ± 0%  0.24µs ± 0%  -79.90%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=false/with-tombstone=true-10     424ns ± 0%   179ns ± 0%  -57.89%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=true/with-tombstone=false-10     432ns ± 1%   441ns ± 0%   +2.23%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=true/with-tombstone=true-10      107ns ± 0%   105ns ± 0%   -1.52%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=false/with-tombstone=false-10   218ns ± 1%   215ns ± 1%   -1.63%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=false/with-tombstone=true-10    176ns ± 0%   174ns ± 0%   -1.03%  (p=0.016 n=4+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=true/with-tombstone=false-10    416ns ± 2%   428ns ± 1%   +2.79%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=true/with-tombstone=true-10     107ns ± 1%   107ns ± 1%     ~     (p=0.952 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=false/with-tombstone=false-10   1.20µs ± 0%  0.24µs ± 0%  -79.63%  (p=0.016 n=4+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=false/with-tombstone=true-10     433ns ± 5%   181ns ± 0%  -58.27%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=true/with-tombstone=false-10     437ns ± 4%   451ns ± 3%     ~     (p=0.056 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=true/with-tombstone=true-10      106ns ± 0%   106ns ± 0%   -0.55%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=false/with-tombstone=false-10   215ns ± 0%   217ns ± 1%     ~     (p=0.135 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=false/with-tombstone=true-10    176ns ± 0%   176ns ± 1%     ~     (p=0.516 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=true/with-tombstone=false-10    418ns ± 1%   442ns ± 2%   +5.72%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=true/with-tombstone=true-10     107ns ± 1%   107ns ± 0%     ~     (p=0.357 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=false/with-tombstone=false-10   1.20µs ± 0%  0.24µs ± 0%  -79.64%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=false/with-tombstone=true-10     425ns ± 0%   181ns ± 1%  -57.38%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=true/with-tombstone=false-10     433ns ± 0%   457ns ± 6%   +5.41%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=true/with-tombstone=true-10      107ns ± 2%   106ns ± 1%     ~     (p=0.159 n=5+5)

name                                            old time/op  new time/op  delta
IteratorSeqSeekGEWithBounds/two-level=false-10   314ns ± 3%   308ns ± 0%  -1.97%  (p=0.008 n=5+5)
IteratorSeqSeekGEWithBounds/two-level=true-10    309ns ± 1%   308ns ± 0%    ~     (p=0.151 n=5+5)
```